日程的增删查改——逻辑判断的补充以及异常的补充

+++ src/main/java/com/program/itta/service/impl/ScheduleServiceImpl.java
@@ -1,6 +1,8 @@
 package com.program.itta.service.impl;
 
 import com.program.itta.common.config.JwtConfig;
+import com.program.itta.common.exception.schedule.ScheduleNameExistsException;
+import com.program.itta.common.exception.schedule.ScheduleNotExistsException;
 import com.program.itta.domain.entity.Schedule;
 import com.program.itta.mapper.ScheduleMapper;
 import com.program.itta.service.ScheduleService;
@@ -23,9 +25,13 @@ public class ScheduleServiceImpl implements ScheduleService {
     private ScheduleMapper scheduleMapper;
     @Resource
     private JwtConfig jwtConfig;
+
     @Override
     public Boolean addSchedule(Schedule schedule) {
-        // TODO Schedule的判断
+        Boolean judgeScheduleExists = judgeScheduleName(schedule);
+        if (judgeScheduleExists) {
+            throw new ScheduleNameExistsException("该任务名称已存在");
+        }
         int insert = scheduleMapper.insert(schedule);
         if (insert != 0) {
             return true;
@@ -35,6 +41,10 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public Boolean deleteSchedule(Schedule schedule) {
+        Boolean judgeScheduleExists = judgeScheduleExists(schedule);
+        if (!judgeScheduleExists) {
+            throw new ScheduleNotExistsException("该任务不存在");
+        }
         int delete = scheduleMapper.deleteByPrimaryKey(schedule.getId());
         if (delete != 0) {
             return true;
@@ -44,6 +54,10 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public Boolean updateSchedule(Schedule schedule) {
+        Boolean judgeScheduleExists = judgeScheduleExists(schedule);
+        if (!judgeScheduleExists) {
+            throw new ScheduleNotExistsException("该任务不存在");
+        }
         int update = scheduleMapper.updateByPrimaryKey(schedule);
         if (update != 0) {
             return true;
@@ -59,5 +73,23 @@ public class ScheduleServiceImpl implements ScheduleService {
             return scheduleList;
         }
         return null;
+    }
+
+    private Boolean judgeScheduleName(Schedule schedule) {
+        List<Schedule> scheduleList = selectByUserId();
+        for (Schedule schedule1 : scheduleList) {
+            if (schedule.getName().equals(schedule1.getName())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private Boolean judgeScheduleExists(Schedule schedule) {
+        Schedule select = scheduleMapper.selectByPrimaryKey(schedule.getId());
+        if (select != null) {
+            return true;
+        }
+        return false;
     }
 }
